+++ src/main/java/com/nebrija/pookemon/model/tipo/Aire.java
@@ -6,11 +6,11 @@ public class Aire extends Pokemon {
 
     public Aire() {}
 
-    public Aire(int id, String nombre, int hp, String habilidad, int ataque, int ataqueEspecial, double defensa, double defensaEspecial, int velocidad, int dodgeProb, Tipo strongAgainst, Tipo weakAgainst, Tipo synergyWith) {
+    public Aire(int id, String nombre, int hp, String habilidad, int ataque, int ataqueEspecial, double defensa, double defensaEspecial, int velocidad, double dodgeProb, Tipo strongAgainst, Tipo weakAgainst, Tipo synergyWith) {
         super(id, nombre, hp, habilidad, ataque, ataqueEspecial, defensa, defensaEspecial, velocidad, dodgeProb, strongAgainst, weakAgainst, synergyWith);
     }
 
-    public Aire(String nombre, int hp, String habilidad, int ataque, int ataqueEspecial, double defensa, double defensaEspecial, int velocidad, int dodgeProb, Tipo strongAgainst, Tipo weakAgainst, Tipo synergyWith) {
+    public Aire(String nombre, int hp, String habilidad, int ataque, int ataqueEspecial, double defensa, double defensaEspecial, int velocidad, double dodgeProb, Tipo strongAgainst, Tipo weakAgainst, Tipo synergyWith) {
         super(nombre, hp, habilidad, ataque, ataqueEspecial, defensa, defensaEspecial, velocidad, dodgeProb, strongAgainst, weakAgainst, synergyWith);
     }
 }
